Added new types of HTTP responses to controllers and made few modifications to the service implementations

+++ src/main/java/pl/jakub/musiclibrary/album/AlbumController.java
@@ -6,6 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/albums")
@@ -19,22 +20,42 @@ public class AlbumController {
 
     @GetMapping
     public ResponseEntity<List<Album>> getAll() {
-        return new ResponseEntity<>(albumService.findAll(), HttpStatus.OK);
+        List<Album> allAlbums = albumService.findAll();
+        if (allAlbums == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (allAlbums.isEmpty())
+            return new ResponseEntity<>(allAlbums, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(allAlbums, HttpStatus.OK);
     }
 
     @GetMapping("/page")
     public ResponseEntity<List<Album>> getSortedPage(@RequestParam Integer pageNumber, @RequestParam Integer pageSize, @RequestParam String sort, @RequestParam(defaultValue = "false") Boolean descending) {
-        return new ResponseEntity<>(albumService.findSortedPage(pageNumber, pageSize, sort, descending), HttpStatus.OK);
+        List<Album> page = albumService.findSortedPage(pageNumber, pageSize, sort, descending);
+        if (page == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (page.isEmpty())
+            return new ResponseEntity<>(page, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(page, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Album> getById(@PathVariable Long id) {
-        return new ResponseEntity<>(albumService.findById(id), HttpStatus.OK);
+        return albumService.findById(id)
+                .map(album -> new ResponseEntity<>(album, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @GetMapping("/by-name")
     public ResponseEntity<List<Album>> getByName(@RequestParam String name) {
-        return new ResponseEntity<>(albumService.findByName(name), HttpStatus.OK);
+        List<Album> albums = albumService.findByName(name);
+        if (albums == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (albums.isEmpty())
+            return new ResponseEntity<>(albums, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(albums, HttpStatus.OK);
     }
 
     @PostMapping
@@ -44,12 +65,21 @@ public class AlbumController {
 
     @PutMapping("/{id}")
     public ResponseEntity<Album> update(@PathVariable Long id, @Valid @RequestBody Album updatedAlbum) {
-        return new ResponseEntity<>(albumService.update(id, updatedAlbum), HttpStatus.OK);
+        Optional<Album> album = albumService.findById(id);
+        if (album.isEmpty())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else
+            return new ResponseEntity<>(albumService.update(id, updatedAlbum), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> delete(@PathVariable Long id) {
-        albumService.delete(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        Optional<Album> album = albumService.findById(id);
+        if (album.isEmpty())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else {
+            albumService.delete(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 }
