+++ src/main/java/pl/jakub/musiclibrary/track/TrackController.java
@@ -6,6 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/tracks")
@@ -19,22 +20,42 @@ public class TrackController {
 
     @GetMapping
     public ResponseEntity<List<Track>> getAll() {
-        return new ResponseEntity<>(trackService.findAll(), HttpStatus.OK);
+        List<Track> tracks = trackService.findAll();
+        if (tracks == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (tracks.isEmpty())
+            return new ResponseEntity<>(tracks, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(tracks, HttpStatus.OK);
     }
 
     @GetMapping("/page")
     public ResponseEntity<List<Track>> getSortedPage(@RequestParam Integer pageNumber, @RequestParam Integer pageSize, @RequestParam String sort, @RequestParam(defaultValue = "false") Boolean descending) {
-        return new ResponseEntity<>(trackService.findSortedPage(pageNumber, pageSize, sort, descending), HttpStatus.OK);
+        List<Track> page = trackService.findSortedPage(pageNumber, pageSize, sort, descending);
+        if (page == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (page.isEmpty())
+            return new ResponseEntity<>(page, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(page, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Track> getById(@PathVariable Long id) {
-        return new ResponseEntity<>(trackService.findById(id), HttpStatus.OK);
+        return trackService.findById(id)
+                .map(track -> new ResponseEntity(track, HttpStatus.OK))
+                .orElseGet(() -> new ResponseEntity(HttpStatus.NOT_FOUND));
     }
 
     @GetMapping("/by-name")
     public ResponseEntity<List<Track>> getByName(@RequestParam String name) {
-        return new ResponseEntity<>(trackService.findByName(name), HttpStatus.OK);
+        List<Track> tracks = trackService.findByName(name);
+        if (tracks == null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if (tracks.isEmpty())
+            return new ResponseEntity<>(tracks, HttpStatus.NO_CONTENT);
+        else
+            return new ResponseEntity<>(tracks, HttpStatus.OK);
     }
 
     @PostMapping
@@ -44,12 +65,21 @@ public class TrackController {
 
     @PutMapping("/{id}")
     public ResponseEntity<Track> update(@PathVariable Long id, @Valid @RequestBody Track updatedTrack) {
-        return new ResponseEntity<>(trackService.update(id, updatedTrack), HttpStatus.OK);
+        Optional<Track> track = trackService.findById(id);
+        if (track.isEmpty())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else
+            return new ResponseEntity<>(trackService.update(id, updatedTrack), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> delete(@PathVariable Long id) {
-        trackService.delete(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        Optional<Track> track = trackService.findById(id);
+        if (track.isEmpty())
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else {
+            trackService.delete(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 }
